Everything but regressions
All tables, figures, and stats are being generated. From previous areas
to be worked
out:
1. the arcsine transformed proportions for norming —DP checked this,
not relevant for these analyses

2. uninf/misc F2 ANOVAs don’t match original; F1 is ok! —DP F2 requires
collapsing; issue is not with script or data set— but may be with the
original analysis

3. Check the arcsine transform for the alternative analyses (both
experiments)— Don’t think this matters
4. Figure out what’s up with the counts being off.— DP compared
original coding spreadsheet used in analysis with the one being read in
here. Counts match when handled appropriately, but issue may have been
with originally reported counts